- Some Changes

+++ src/main/kotlin/ConfigurationManager.kt
@@ -1,6 +1,6 @@
 
 data class ConfigurationManager(
-    val frameRate: Double = 30.0,
+    val frameRate: Int = 30,
     val width: Int = 1280,
     val height: Int = 720,
     val videoCodecName: String = "libx264",
